requirements.txt
deepdiff
pytest
pyyaml
